Update DEEPLABCUT.yaml
Users appear to be grabbing `DEEPLABCUT.yaml` from the repo in order to install `2.2rc3`. Unfortunately, it installs 2.1.10.4, and gets tables from conda rather than pip. Therefore, the instructions elsewhere to install the newest versions of the code with this conda env file are inaccurate. 

To eliminate confusion, the repo version of `DEEPLABCUT.yaml` should be updated to the same as the version on the blog.

On windows 10, I have tried using the existing repo file to install 2.1.10.4, then running `pip install deeplabcut[gui]==2.2rc3`. While it appears to install ok, I get a tables error (seen in #1368 and #1398 ). Uninstalling pytables with conda and reinstalling with pip works sometimes (#1368), but just now I tested it again and created a tensorflow error.