DOC: Include API documentation in jupyter book docs
At the moment, the API documentation seems to be hosted manually
in the GitHub wiki on the DeepLabCut repository. Links to sections
of the wiki are provided in the jupyter book docs. This is not ideal
as the wiki needs to be maintained manually every time changes are
made to the docstrings.

This is a proof of concept to include the api documentation in the
jupyter book docs directly. This is achieved using the special
eval-rst directive and by manually including an rst file in the
docs. The included rst file is manually created to document the
function.

Additionally, the numpydoc extension is included to conform to
existing standards in the ecosystem and make the docs easier to
read.

In order to build the jupyter book docs now, the deeplabcut package
needs to be installed - and the numpydoc package.

	modified:   _config.yml
	modified:   deeplabcut/create_project/new.py
	new file:   docs/api/deeplabcut.create_project.new.rst
	modified:   docs/standardDeepLabCut_UserGuide.md