CLN: Rename api docs files
the api docs files have been renamed for ease of understanding.
earlier, the filenames were the paths to the module which contained
the function. now, the filenames are just deeplabcut.function_name.

note that this is also done as multiple functions defined in
the same module needed to be documented separately

	renamed:    docs/api/deeplabcut.generate_training_dataset.trainingsetmanipulation.rst -> docs/api/deeplabcut.check_labels.rst
	renamed:    docs/api/deeplabcut.create_project.new.rst -> docs/api/deeplabcut.create_new_project.rst
	renamed:    docs/api/deeplabcut.generate_training_dataset.frame_extraction.rst -> docs/api/deeplabcut.extract_frames.rst
	modified:   docs/standardDeepLabCut_UserGuide.md