REF: Denest filterpredictions for loop
there were two try/excepts, one within the other.
this commit simplifies the code by pulling the inner try/except out.

the code now tries to look for filtered data. if filtered data is found
for the video, we continue to the next item in the loop.

if not found, we catch and ignore the error. we then try to load the
data and filter it

	modified:   deeplabcut/post_processing/filtering.py