REF: Reduce the code within the try block
the second try block in filterpredictions looked for the analyzed file
and performed operations to filter it - and if the file is not found,
continue to the next item in the for loop.

this commit reduces the try block to only load the analyzed file and
immediately catch the error and continue if the file doesnt exist.

the rest of the filtering on the analyzed data is performed outside the
try/except block.

	modified:   deeplabcut/post_processing/filtering.py