Address underlying compatibility issue with scipy 1.11.0
See warning in scipy 1.9's documentation.
See warning in https://docs.scipy.org/doc/scipy-1.9.0/reference/generated/scipy.stats.mode.html

The SPEC0 suggests that scipy 1.8 be dropped in 2024, but I think it is
OK to drop it a little early.

If you don't want to drop it, we could add a version check for the
keyword argument, but that seems a little much.

The keyword argument `keepdims` was added in scipy 1.9

xref: https://github.com/DeepLabCut/DeepLabCut/pull/2290