Update evaluate_multianimal.py
Fix pandas FutureWarning in `evaluate_multianimal.py` related to `inplace=True` in `.replace()`

This Pull Request addresses a `FutureWarning` originating from the pandas library that appears during the evaluation of multi-animal pose estimation models in DeepLabCut.

**Problem:**

The warning is triggered by the use of the `.replace()` method with the `inplace=True` argument in the `evaluate_multianimal.py` file (likely around line [**Insert the line number you found, e.g., 346**]). Pandas is discouraging the use of `inplace=True` in combination with chained assignment, as it can lead to unexpected behavior and will be removed in a future version.

**Solution:**

This PR modifies the line to remove the `inplace=True` argument and instead assigns the result of the `.replace()` operation back to the DataFrame column. For example, the line was changed from:

```python
temp["bodyparts"].replace(..., inplace=True)