Fix ambivalous interpretation of cropping=None in `analyze_videos`
In the original behaviour, cropping=None was interpreted as read cropping params from config. This is in contrast to most other occurrences (`video_inference`, `gui.tabs.analyze_videos`, `VideoIterator`), which interpret cropping=None as no cropping. This leads to potential conflicts.

This commit changes the behaviour for cropping=None, which is now interpreted as no cropping (like elsewhere).

For backwards compatibility, the default argument is changed to 'from_config' which follows prior behavior.