Fix bug misplaced bounding boxes when `displaycropped=True`.
Bounding box positions were corrected by the crop margin, even when the output video is also cropped (which should not be the case). This leads to misplaced bounding boxes.

This commit fixes the unintended behaviour where bounding box positions are always corrected: i.e. bounding boxes are only corrected when cropping was used during video AND the output video is not cropped.